[SDPL-96] Split RplSelect into its own component
Split fieldRplSelect into its own RplSelect component so that it can be used independently of Vue Form Generator
