+++ packages/components/Molecules/Form/stories.js
@@ -1,5 +1,40 @@
 import { storiesOf } from '@storybook/vue'
 import RplForm from './index.vue'
+import RplSelect from './Select.vue'
+
+storiesOf('Molecules/Form/Select', module)
+  .add('Single', () => ({
+    components: { RplSelect },
+    template: `<rpl-select :values="values" :config="config"></rpl-select>`,
+    data () {
+      return {
+        config: {
+          multiselect: false,
+          placeholder: 'Select',
+          showitems: 4,
+          fieldId: 'select',
+          label: ''
+        },
+        values: [{ id: 'vic', name: 'Vic' }, { id: 'nsw', name: 'New South Wales' }, { id: 'wa', name: 'Western Australia' }, { id: 'ql', name: 'Queensland' }]
+      }
+    }
+  }))
+  .add('Multiselect', () => ({
+    components: { RplSelect },
+    template: `<rpl-select :values="values" :config="config"></rpl-select>`,
+    data () {
+      return {
+        config: {
+          multiselect: true,
+          placeholder: 'Select',
+          showitems: 4,
+          fieldId: 'select',
+          label: ''
+        },
+        values: [{ id: 'vic', name: 'Vic' }, { id: 'nsw', name: 'New South Wales' }, { id: 'wa', name: 'Western Australia' }, { id: 'ql', name: 'Queensland' }]
+      }
+    }
+  }))
 
 storiesOf('Molecules/Form', module)
   .add('Default', () => ({
